More factoring to support not calling properties off elements directly
* separates accessors from patching
* logical-mutation renamed to methods-mutation
